+++ lib/tmuxinator/hooks.rb
@@ -2,7 +2,8 @@ module Tmuxinator
   module Hooks
     module_function
 
-    def commands_from(hook_config)
+    def commands_from(project, hook_name)
+      hook_config = project.yaml[hook_name]
       if hook_config.is_a?(Array)
         hook_config.join("; ")
       else