+++ CHANGELOG.md
@@ -5,7 +5,7 @@
 - Add tmux 2.2 and 2.3 the TravisCI test matrix
 
 ### New Features
-- add on_start, on_detach, on_reattach and on_stop hooks for project
+- add on_project_start, on_project_first_start, on_project_restart, on_project_exit and on_project_stop hooks for project
 
 ## 0.9.0
 ### Misc